web-audio/no-prefix.html

<!doctype html>
<html class="a">
 <head>
  <title>Vendor Prefixes</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/lodash.js"></script>
 </head>
 <body class="a">
  <div id="log"></div>
  <script>

function assert_defined(obj, desc) {
  assert_true(undefined != obj, desc);
}

var context;
setup(function() {
  context = new AudioContext();
});

test(function() {
  assert_defined(AudioContext, "AudioContext is defined")
}, "prefix");

  </script>
 </body>
</html>
